Move `TypeBinding` class to Util folder.
