+++ TrackDesigner/MainFormViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
+using TrackDesigner.Util;
 
 namespace TrackDesigner;
