+++ TrackDesigner/MainWindow.xaml.cs
@@ -1,5 +1,4 @@
-﻿using System.Runtime.CompilerServices;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls.Ribbon;
 
 namespace TrackDesigner
